+++ cypress/integration/notes-list.spec.js
@@ -86,7 +86,7 @@ context('Notes list', () => {
 
       cy.wait(1000)
 
-      cy.get('#main-search').type('link')
+      cy.get('#main-search').find('input').type('link', { force: true })
 
       cy.wait('@linkAutocomplete')
 
@@ -228,7 +228,7 @@ context('Notes list', () => {
 
       cy.wait(1000)
 
-      cy.get('#main-search').type('link')
+      cy.get('#main-search').find('input').type('link', { force: true })
 
       cy.wait('@linkAutocomplete')
 
@@ -360,7 +360,7 @@ context('Notes list', () => {
 
       cy.wait(1000)
 
-      cy.get('#main-search').type('link')
+      cy.get('#main-search').find('input').type('link', { force: true })
 
       cy.wait('@linkAutocomplete')